Nicolas/514 676 enable one click deploy of our internal moose (#686)
Adds Workflow to automatically update our moose project when moose gets updated
Adds integration between 1password and github actions to be able to use secrets from 514's 1password
Adds integration between Github Action and github to be able to use github actions for continuous delivery
Fixes an issue when the versions directory was not created and a build was run
Changes the Docker Hub credentials to use one from a service account
Adds a build command to the project template when running init

+++ apps/framework-cli/src/cli/routines/clean.rs
@@ -149,19 +149,26 @@ impl DeleteVersions {
 impl Routine for DeleteVersions {
     fn run_silent(&self) -> Result<RoutineSuccess, RoutineFailure> {
         let versions_dir = self.internal_dir.join("versions");
-        fs::remove_dir_all(&versions_dir).map_err(|err| {
-            RoutineFailure::new(
-                Message::new(
-                    "Failed".to_string(),
-                    format!("to remove versions directory at {}", versions_dir.display()),
-                ),
-                err,
-            )
-        })?;
+        if !versions_dir.exists() {
+            return Ok(RoutineSuccess::success(Message::new(
+                "ok".to_string(),
+                "No versions directory to remove".to_string(),
+            )));
+        } else {
+            fs::remove_dir_all(&versions_dir).map_err(|err| {
+                RoutineFailure::new(
+                    Message::new(
+                        "Failed".to_string(),
+                        format!("to remove versions directory at {}", versions_dir.display()),
+                    ),
+                    err,
+                )
+            })?;
 
-        Ok(RoutineSuccess::success(Message::new(
-            "Successfully".to_string(),
-            "removed versions directory".to_string(),
-        )))
+            Ok(RoutineSuccess::success(Message::new(
+                "Successfully".to_string(),
+                "removed versions directory".to_string(),
+            )))
+        }
     }
 }
